[FE/refactor] home 폴더 리팩토링
