+++ frontend/lib/screens/main/bottom_tabs/home/home_state_box_widgets/home_challenge_state_box.dart
@@ -40,6 +40,7 @@ class _ChallengeStateBoxState extends State<ChallengeStateBox> {
       challengeSimples.clear();
       challengeSimples.addAll(
           controller.myChallenges.where((c) => c.status == "진행중").toList());
+
     });
   }
